~more deprecations and cleanups
